Enhance SHOW STORAGE INFO query with storage_mode
